requirements.txt
sentence-transformers
numpy
pinecone
python-dotenv
fastapi
uvicorn
APScheduler
gunicorn